requirements/base.txt

# Django
django==2.0.3
djangorestframework==3.8.2
django_filter==1.1.0
dj_database_url==0.4.2
djangorestframework-jwt==1.11.0
django-allauth==0.36.0
django-rest-auth==0.9.3

psycopg2-binary==2.7.4

whitenoise==3.3.1

# Raven
raven==5.22.0

# Misc
python-decouple==3.1
requests==2.18.4
django-widget-tweaks==1.4.2

# Celery
celery==4.1.0
django-celery-beat==1.1.1
django-celery-results==1.0.1
flower==0.9.2

# Django Webpack
django-webpack-loader==0.6.0

# Time Zones
pytz==2016.10
